Move settings_changer to _lib
This is more correct location for launcher specific tools.
